+++ logisland-components/logisland-processors/logisland-processor-enrichment/src/main/java/com/hurence/logisland/processor/enrichment/IpToGeo.java
@@ -16,6 +16,7 @@
 package com.hurence.logisland.processor.enrichment;
 
 import com.hurence.logisland.annotation.documentation.CapabilityDescription;
+import com.hurence.logisland.annotation.documentation.ExtraDetailFile;
 import com.hurence.logisland.annotation.documentation.Tags;
 import com.hurence.logisland.classloading.PluginProxy;
 import com.hurence.logisland.component.PropertyDescriptor;
@@ -47,6 +48,7 @@ import static com.hurence.logisland.service.iptogeo.IpToGeoService.*;
         " one must define a comma separated list of fields as a value for the **" + IpToGeo.PROP_GEO_FIELDS + "** property. The list of the available geo fields" +
         " is in the description of the **" + IpToGeo.PROP_GEO_FIELDS + "** property."
 )
+@ExtraDetailFile("./details/IpToGeo-Detail.rst")
 public class IpToGeo extends IpAbstractProcessor {
 
     private static Logger logger = LoggerFactory.getLogger(IpToGeo.class);