+++ logisland-documentation/changes.rst
@@ -18,6 +18,9 @@ v1.0.0
 - add Cassandra datastore service
 - support all Kafka connect sinks
 - add KafkaStreams engine
+- update documentation
+- fix test framework (runner)
+- added vanilla java engine
 
 
 v0.14.0